requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.5
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.2
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
kombu==5.3.4
packaging==23.2
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.10
